Locking: make the owner optional - for Office2003
Office 2003 sends out an incorrectly formed lockinfo declaration:

<?xml version="1.0" encoding="UTF-8" ?>
<lockinfo xmlns="DAV:">
<locktype>
<write/>
</locktype>
<lockscope>
<exclusive/>
</lockscope>
<owner>Administrator</owner>
</lockinfo>

because the <owner> tag does not contain an <href> but rather the
user name.

Moreover, if the LOCK response contains a properly structured owner
information - with an <href> tag - Office2003 will discard it and
mark the document as read only.

We could support it by rendering a wrong response, but as Office2003
works correctly if there is no owner at all in the lockinfo reply,
we chose to pursue that path.