Fix LOCK crash on Nokogiri >= 1.6
Nokogiri >= 1.6 raises RuntimeError when given an empty string as
document input with "strict" config, whilst 1.5 just returned an
empty Nokogiri::XML::Document.

This patch also adds a broad rescue on RuntimeError, because
Nokogiri raises that generic exception if it cannot parse the
document. This should be better handled with Nokogiri maintainers.